tests: use portable 64-bit division types
The signed and unsigned 64-bit division sources should use portable C spellings while retaining the existing focused coverage. Address R1#26 by using long long and unsigned long long in the s64-div and longlong-udiv fixtures.

+++ tests/end_to_end/s64-div/orig.c
@@ -1,1 +1,1 @@
-__int64 test(__int64 a, __int64 b) { return a / b; }
+long long test(long long a, long long b) { return a / b; }
